+++ src/OpenTelemetry.ResourceDetectors.AWS/CHANGELOG.md
@@ -6,7 +6,7 @@
   AWS ECS detector.
   ([#1552](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1552))
 * Implement support for cloud.resource_id attribute in AWS ECS detector.
-  ([#???](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/???))
+  ([#1576](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1576))
 
 ## 1.4.0-beta.1